+++ packages/website/src/_services/maps/map-bubble/map-bubble.ts
@@ -12,6 +12,7 @@ const DATA_LAYER_ID = 'species-information'
 const LABEL_LAYER_IDS = ['tunnel-primary-secondary-tertiary-case', 'tunnel-major-link-case', 'tunnel-motorway-trunk-case', 'tunnel-path', 'tunnel-steps', 'tunnel-major-link', 'tunnel-pedestrian', 'tunnel-primary-secondary-tertiary', 'tunnel-oneway-arrow-blue', 'tunnel-motorway-trunk', 'tunnel-oneway-arrow-white', 'ferry', 'ferry-auto', 'road-pedestrian-case', 'road-street-low', 'road-street-case', 'road-secondary-tertiary-case', 'road-primary-case', 'road-major-link-case', 'road-motorway-trunk-case', 'road-path', 'road-steps', 'road-major-link', 'road-pedestrian', 'road-street', 'road-secondary-tertiary', 'road-primary', 'road-oneway-arrow-blue', 'road-motorway-trunk', 'road-oneway-arrow-white', 'bridge-pedestrian-case', 'bridge-primary-secondary-tertiary-case', 'bridge-major-link-case', 'bridge-motorway-trunk-case', 'bridge-path', 'bridge-steps', 'bridge-major-link', 'bridge-pedestrian', 'bridge-primary-secondary-tertiary', 'bridge-oneway-arrow-blue', 'bridge-motorway-trunk', 'bridge-major-link-2-case', 'bridge-motorway-trunk-2-case', 'bridge-major-link-2', 'bridge-motorway-trunk-2', 'bridge-oneway-arrow-white', 'aerialway', 'admin-1-boundary-bg', 'admin-0-boundary-bg', 'admin-1-boundary', 'admin-0-boundary', 'admin-0-boundary-disputed', 'road-label', 'road-number-shield', 'road-exit-shield', 'waterway-label', 'natural-line-label', 'natural-point-label', 'water-line-label', 'water-point-label', 'poi-label', 'transit-label', 'airport-label', 'settlement-subdivision-label', 'settlement-label', 'state-label', 'country-label']
 
 const DEFAULT_RADIUS = 5
+const DEFAULT_BOOLEAN_RADIUS = 2
 
 export default class MapBubbleComponent extends Vue {
   @Prop() mapId!: string
@@ -83,7 +84,7 @@ export default class MapBubbleComponent extends Vue {
   getRadius (datum: MapSiteData): number {
     const value = datum.distinctSpecies[this.dataKey]
     const { maximumRadius, dataRange } = this.dataset
-    if (dataRange) {
+    if (dataRange && (typeof value !== 'boolean')) {
       const range = dataRange[this.dataKey]
       const radius = maximumRadius ?? range.length
       const belongRange = range.findIndex(r => value <= r)
@@ -91,11 +92,20 @@ export default class MapBubbleComponent extends Vue {
       return (belongRange + 1) * ratio
     }
 
-    return (typeof value === 'boolean') ? DEFAULT_RADIUS : Math.sqrt(value)
+    return (typeof value === 'boolean') ? DEFAULT_BOOLEAN_RADIUS : Math.sqrt(value)
   }
 
   getPopup (datum: MapSiteData): string {
     return this.getPopupHtml(datum, this.dataKey)
+  }
+
+  getColor (datum: MapSiteData): string {
+    const value = datum.distinctSpecies[this.dataKey]
+    if (typeof value !== 'boolean') {
+      return this.dataset.color
+    }
+
+    return value ? this.dataset.color : '#CFCFCF'
   }
 
   setupMapPopup (): void {
@@ -149,7 +159,8 @@ export default class MapBubbleComponent extends Vue {
         properties: {
           title: datum.siteName,
           radius: this.getRadius(datum),
-          popup: this.getPopup(datum)
+          popup: this.getPopup(datum),
+          color: this.getColor(datum)
         }
       })
       )
@@ -170,7 +181,7 @@ export default class MapBubbleComponent extends Vue {
         source: DATA_LAYER_ID,
         paint: {
           'circle-radius': ['*', ['get', 'radius'], 4], // TODO 41 - Normalize circle size
-          'circle-color': this.dataset.color || '#B42222',
+          'circle-color': ['string', ['get', 'color'], '#CFCFCF'],
           'circle-stroke-width': 0.1,
           'circle-opacity': 0.6
         }