+++ packages/website/src/activity-patterns/components/activity-patterns-by-location/activity-patterns-by-location.vue
@@ -2,10 +2,10 @@
   <div
     class="w-full"
   >
-    <div class="flex flex-row justify-between  items-center">
+    <div class="flex flex-row justify-between items-center">
       <div class="flex flex-row items-center">
         <select
-          v-model="selectedType"
+          v-model="selectedDetectionType"
           class="text-xl py-1 bg-mirage-grey border-t-0 border-l-0 border-r-0 border-b-2 focus:(border-box-grey border-t-0 border-l-0 border-r-0 border-b-2 ring-0 outline-none)"
         >
           <option
@@ -33,12 +33,34 @@
         v-for="(dataset, idx) in datasets"
         :key="idx"
         :dataset="dataset"
-        :data-key="selectedType"
+        :data-key="selectedDetectionType"
         :get-popup-html="getPopupHtml"
-        :map-id="`activity-patterns-by-location-${idx}`"
+        :map-id="`activity-patterns-detection-by-location-${idx}`"
         :map-config="config"
         :map-style="mapStyle"
-        :map-export-name="mapExportName(dataset)"
+        :map-export-name="mapExportName(dataset, selectedDetectionType)"
+        :is-show-labels="isShowLabels"
+        class="w-full"
+        @emit-map-moved="mapMoved"
+      />
+    </div>
+    <h2 class="text-white text-xl mt-5 ml-1">
+      Occupancy by site
+    </h2>
+    <div
+      class="grid gap-2 mt-2"
+      :class="{ [`md:grid-cols-${columnCount}`]: true }"
+    >
+      <map-bubble-component
+        v-for="(dataset, idx) in datasets"
+        :key="idx"
+        :dataset="dataset"
+        :data-key="occupancyType"
+        :get-popup-html="getPopupHtml"
+        :map-id="`activity-patterns-occupancy-by-location-${idx}`"
+        :map-config="config"
+        :map-style="mapStyle"
+        :map-export-name="mapExportName(dataset, occupancyType)"
         :is-show-labels="isShowLabels"
         class="w-full"
         @emit-map-moved="mapMoved"